requirements/development.txt
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.19.0
